Completed on the code

+++ src/main/java/com/rediscache/redis/services/ProductService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ProductService {
     public Product saveProduct(Product product);
-    public Product updateProduct(Product product, Integer productId);
-    public void deleteProduct(Integer productId);
-    public Product getProductById(Integer productId);
+    public Product updateProduct(Product product, Long productId);
+    public void deleteProduct(Long productId);
+    public Product getProductById(Long productId);
     public List<Product> getAllProducts();
 }
